backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djoser==2.3.3
django-cors-headers==4.7.0
psycopg2
